task fixes

+++ bin/setup/tasks/mail.sh
@@ -5,8 +5,8 @@ echo -e "\nSetup msmtp for root ..."
   && echo 'skipped, requires root' \
   && exit 1
 
-echo -e "... install msmtp"
 ! command -v msmtp &> /dev/null \
+  && echo "... install msmtp" \
   && ! aptitude install msmtp msmtp-mta \
   && echo "failed install" && exit 1
 
@@ -15,17 +15,18 @@ echo "aliases /etc/aliases" | tee /etc/msmtprc >/dev/null
 chmod 600 /etc/msmtprc
 echo "TODO add mails to /etc/aliases"
 
-echo -e "... setup user mailing"
 ## FIXME password is exposed to ps in sed command
-[ ! -f ~/.msmtprc ] \
-  && read -sp "Mailing password: " password && echo \
-  && [ ! -z "$password" ] \
-  && cp /opt/msladek/stuff/etc/msmtprc ~/.msmtprc \
-  && chmod 600 ~/.msmtprc \
-  && sed -i -e "s/<hostname>/$(hostname)/g" ~/.msmtprc \
-  && sed -i -e "s/<password>/${password}/g" ~/.msmtprc \
-  && -e "Subject: Test MSMTP\n\norigin: $(hostname)" | msmtp ${USER}@sladek.co \
-  || echo 'unable to setup ~/.msmtprc'
+if [ ! -f ~/.msmtprc ]; then
+  echo -e "... setup user mailing"
+  read -sp "Mailing password: " password && echo
+  [ ! -z "$password" ] \
+    && cp /opt/msladek/stuff/etc/msmtprc ~/.msmtprc \
+    && chmod 600 ~/.msmtprc \
+    && sed -i -e "s/<hostname>/$(hostname)/g" ~/.msmtprc \
+    && sed -i -e "s/<password>/${password}/g" ~/.msmtprc \
+    && -e "Subject: Test MSMTP\n\norigin: $(hostname)" | msmtp ${USER}@sladek.co \
+    || echo 'unable to setup ~/.msmtprc'
+fi
 unset password
 
 exit 0
